chore(rust): update snapshots
Updated struct and enum fields in Rust model and SDK test snapshots to use i64 instead of i32 for integer types. Also added PartialEq, Eq, and Hash derives to several types for improved comparison and hashing. Fixed Option<String> mapping in SDK error parsing logic to avoid unnecessary allocation.

+++ generators/rust/model/src/__test__/snapshots/union-types/types_car.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub struct Car {
-    pub doors: i32,
+    pub doors: i64,
     pub fuel_type: String,
 }
